+++ smart-contracts/wasm-chain-integration/src/v1/crypto_primitives_tests.rs
@@ -50,9 +50,7 @@ where
 fn test_crypto_prims() -> anyhow::Result<()> {
     let nrg = 1_000_000_000;
 
-    let start_energy = InterpreterEnergy {
-        energy: nrg * 1000,
-    };
+    let start_energy = InterpreterEnergy::new(nrg * 1000);
 
     let skeleton = parse::parse_skeleton(CONTRACT_BYTES).unwrap();
     let module = {
@@ -60,6 +58,7 @@ fn test_crypto_prims() -> anyhow::Result<()> {
             ValidationConfig::V1,
             &ConcordiumAllowedImports {
                 support_upgrade: true,
+                enable_debug:    false,
             },
             &skeleton,
         )
@@ -113,7 +112,7 @@ fn test_crypto_prims() -> anyhow::Result<()> {
         };
         let inner = mutable_state.get_inner(&mut backing_store);
         let state = InstanceState::new(backing_store, inner);
-        let mut host = ReceiveHost::<_, Vec<u8>, _> {
+        let mut host = ReceiveHost::<_, Vec<u8>, _, _> {
             energy: start_energy,
             stateless: StateLessReceiveHost {
                 activation_frames: MAX_ACTIVATION_FRAMES,
@@ -124,6 +123,7 @@ fn test_crypto_prims() -> anyhow::Result<()> {
                 params: super::ReceiveParams::new_p5(),
             },
             state,
+            trace: (),
         };
         let r = artifact.run(&mut host, name, args);
         match r {